tests/migrations: make params JSON-friendly
This is to launch tests with individual parameter combinations.

Ducktape needs parameter types to be JSON-serializable using default
seralizer, and also when de-serialized again to be equal to the original
value.
The way it deserializes, only JSON objects are supported, JSON arrays
will morph into sets of parameter combinations. So we need the type to
be dict-like: serialize the same way and be equal to a dict too. Default
serializer will only consider actual dicts.

It would be ideal to create a pseudo-metaclass akin
TypedDict/NamedTuple, such that,
- like with these two, classes are defined by type-hinted slots;
- constructor accepts unnamed arguments, like NamedTuple;
- objects constructed are actual dicts, like TypedDict creates;
- classes to expose enum types of its slots.
Unfortunately NamedTuple and TypedDict themselves won't easily extend,
so I'd have to mock them.

So at least for now use TypedDict.