rpc: Change max internal rpc connection count limit
Previously we limited this to 32 for fear of connection explosion on
larger core machines but we haven't seen any issues so far and 64 core
machines (e.g.: AWS 16xlarge) are becoming more common so up the default
to account for this.

Having one per shard is important for even load distribution across
shards.