dt/config_test: Add OIDC_ALLOW_LIST
Fixes CORE-7816

Allow logs of the form 'security - .* - Error updating...', which will appear
when the configured OIDC endpoint is not available.

This is benign in cases like the test for valid cluster config settings,
where OIDC _functionality_ is not the subject of the test.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>
(cherry picked from commit 1dce16a77500f3963fd50dfda61819c598b28715)