iceberg: create namespace on create_table not found error
Updates the rest_catalog implemenetaiton to respond to not_found errors
on table creation. These are presumably seen when the namespace does not
exist. In the face of such an error, the catalog will attempt to create
the namespace and then try table creation again.

This commit updates a ducktape test to expect this, though AFAICT this
isn't exercising this new codepath yet since ducktape isn't wrapping a
catalog impl that requires namespaces when creating a table.

I have manually validated that that this change allows us to operate
against Open Catalog without an separate explicit namespace creation
step.