dt: Allow selecting key type for certificates in DT tests
This change permits users of the TLSCertManager to select which type of
key to use in certificates.  If one isn't selected, a random one is
chosen, in the hope this increases our test coverage over time.

Additionally, the tls_version_test has been updated to test both ECDSA
and RSA based certificates.

Signed-off-by: Michael Boquard <michael@redpanda.com>