iceberg: add basic retries for metadata io
I noticed that in prolonged tests, we would eventually hit EOS errors
very reliably, particularly when downloading many manifests in a row as
a part of a merge append. I don't fully understand why the errors were
so common, but a retry policy at least helps us not get stuck.

I'd originally wanted to build a more robust retry policy, but this
will do for now.