cloud_storage: Allow remote_segment concurrency
Previously, the remote_segment was assuming that there is no concurrency
and we are creating only one instance of it per segment. The
'remote_segment' depends on cloud storage cache which acts as a shared
global state for all 'remote_segment' instances. If two segments are
created in parallel there is a slight chance that they could interact
through this global state. If one segment starts download the cache will
be in the 'in-progress' state for this key. Then another segment may try
to trigger download and see the 'in-progress' state and trigger
assertion.

Normally, we're not creating two instancess of the same segment in TS.
But there is a slight chance that under very high contention when the
materialized segments are churning rapidly we will evict the segment and
immediately recreate it. If both of them are in the right state they may
trigger the assertion. The segments are evicted by the
'remote_partition' which moves them to the eviction queue. So basically,
the segment may exist in the eviction queue for very short time. But
under hight contention it could be enough.

This commit fixes this issue by turning the assertion into the
exception. The exception is propagated to the 'data_stream' call and
then to the client.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit c2039ddf3997dccde6a123339152582ccbb61e52)