s/log: fixed race condition between appends and prefix truncation
The race condition between prefix truncation and appends was leading to
an issue in which the same offset might have been assigned twiice to
batches that were appended to log. This invalidates the invariant and
may lead to inconsistencies.

The fix has two parts. The first part fixes calculation of next offset
for the newly created log appender. Previously the check incorrectly
validated if the log is empty. The condition was simplified and it is
now based on an invariant that if the `start_offset > dirty_offset` the
log is empty (there is one exception to it when the log is empty and
first initialized, but then the next offset is derived as usual by
incrementing the dirty offset).

The second part of the fix is an addition of check if the segment is
still removable after the segment write lock was acquired for its
removal. It is perfectly viable scenario that when the prefix truncation
fiber is waiting for the lock the segment is written to. In this case
the segment should not be removed as it contains valid data.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 3709f8108036ffa18fdcee51a450fa7621f82f3e)