`storage`: reset compaction window start offset when offset map isn't built
Consider the following situation:

[0] [1] [2] [3] [4] [5]
            ^
	    |----- _last_compaction_window_start_offset

In the event that segment [2] cannot have its keys completely added
to the offset map due to memory constraints, if we do not reset
the window start offset, compaction will become somewhat stuck as it
repeatedly tries to index segment [2] in future compaction rounds.

By resetting the window's value in this situation, we can allow
compaction to continue to make forward progression.