datalake: make purging table idempotent
Previously, if we failed to complete topic purging process after
dropping the table (e.g. if a topic table update failed), the topic
would get stuck in coordinator as the file committer would try to load
non-existent table.

Fix this by returning early if there are no files to commit.