dt/redpanda: increase expect_fail timeout to 40s
This gives a starting redpanda node the same time to fail as we give it
to successfully start up. It also allows reusing the timeout parameter
to configure how much to wait for redpanda to fail on startup (which
some call sites already attempt to do).

Reusing the startup timeout is a better guess than the current static
10s, because tests that expect a timeout typically expect it to happen
during the same timeframe as what a normal startup would take.

This also means that the timeout is increased from the current 10s to
40s (`DEFAULT_NODE_READY_TIMEOUT_SEC`) by default. Which solves some CI
failures on Azure CDT where startup is occasionally slower.