archival: consistent log size probes across replicas
We called update probe only from leaders and after exiting the upload
loop which led to inconsistent and stale metrics.

Fix this by introducing a subscription mechanism to the STM which
is the source of truth for the manifest state and must be consistent
across all replicas.

The first attempt was in
https://github.com/redpanda-data/redpanda/pull/24257 but the feedback
suggested that the approach in this commit is better.