c/controller_backend: allow `shutdown_partition` to fail on app shutdown
With `ss::gate_closed_exception` only. This is to untangle a race
condition between partition shutdown and app shutdown. Callers should
consider the operation unsuccessful and will probably retry it, as
`shard_placement_table` is not going to be updated.

While this unsuccessful operation will leave the partition in a somewhat
orphaned state, i.e. without appropriate `shard_table` entries, the rest
of the system should be able to live with it in the same way as it lives
during a successful partition shutdown.