`storage`: add `chunked_sliding_window_compact()`
In the case that zero segments were able to be indexed for a round of sliding
window compaction, chunked compaction must be performed.

This implementation uses some of the current abstractions from the compaction
utilities to perform several rounds of sliding window compaction with a
partially indexed map created from the un-indexed segment in a linear fashion.

This implementation is sub-optimal for a number of reasons- namely,
that segment indexes are read and rewritten each time a round of chunked
compaction is performed. These intermediate states are then used for the
next round of chunked compaction.

In the future, there may be a more optimal way to perform these steps
using less IO by holding more information in memory before flushing
the final results to disk, and not every intermediate stage.