tests: wrapped querying high watermark with exception handler
When querying high watermark it may happen that the broker responds with
an error in this case the test should retry the check. Previously the
exception was causing the consum loop to stop and the test to fail.

Fixes: CORE-6993

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit c12c430d5fe58ca5faf456ae45af3647ff74116d)