rptest/s: continue waiting for manifest uploads after MaxRetryError
At the end of tests which use tiered storage we do an object storage
scrub which relies on partition manifests to be up to date.

In some tests which otherwise succeed we might not have stable
leadership by the end of the test which results in 503 being thrown from
the status endpoint due to a missing leader. Internally, we retry these
requests a couple of times in a quick succession but this is often not
enough for the partition to have a leader and we end up with failed
test.

This commit catches the MaxRetryError and ignores it thus deferring the
retry/timeout conditions to the caller. I.e. `wait_for_manifest_uploads`
will wait at least 30s which is a much better timeout for leadership to
be established and become stable.