c/leader_balancer: implement and use even_node_load_constraint
Leader balancer treats all CPU cores available in the cluster as independent and
tries to balance leadership among them and not among nodes (i.e. the objective is
that each core has the same number of leaders, and not each node). This
leads to unintuitive results, especially when the number of raft groups
is comparable to the number of available shards.

Add a low-priority node balancing constraint to fix that.

(cherry picked from commit f0a44288ec55d80d27551d6db61956ee02b9863a)