raft: added new strategy to make reconfiguration symetric
In order to support cancellation of raft group reconfiguration
cancellation we must make the cancell operation symmetric. This means
that when reconfiguration is cancelled twice it must lead to the same
outcome as it would not be cancelled at all.

This commit introduces a new `raft::group_configuration` version
together with a new reconfiguration strategy. This is a first step
towards allowing cancelling cancellations in controller.

Signed-off-by: Michał Maślanka <michal@redpanda.com>