http: Shutdown connection if shutdown_now was called
If 'shutdown_now' was called we may still establish the connection. In
this case we're still returning the time_out error but the http
connection remains in wrong state. To fix this this commit adds
'shutdown' call if 'shutdown_now' was called and we're attempting to
exit the connection loop ASAP.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>