ct: Add 'interfaces' submodules
The module is responsible for interop between cloud topics and the rest
of redpanda. Currently, it contains only 'partition_manager' adapter.
The adapter is supposed to be used in cloud topics to find partition
and read data from the underlying raft group. It can also be easily
mocked in tests.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>