archival: calculate time based retention from first addressable segment
apply_retention() and garbage_collect() are 2 distinct routines. The
former computes the new start offset while the later removes the
physical files and removes the segment entries from the manifest.

However, if we bail out of the second routine (e.g. timeout, leadership
lost) we will call apply_retention() again. If we apply the retention
rules starting with the first segment entry in the manifest it can
happen that we reach the same result (i.e. same new start offset). When
this happens, we replicate a new update_start_offset command to the
archival stm which tries to advance_start_offset to the current start
offset. Currently, we log an error in that case from
`archival_metadata_stm::apply_update_start_offset`. Nothing is broken
beside printing that error log line.

```
if (!_manifest->advance_start_offset(so.start_offset)) {
    vlog(
      _logger.error,
      "Can't truncate manifest up to offset {}, offset out of range",
      so.start_offset);
}
```

This commit avoids this annoyance and avoids unnecessary redundant work
for retention computation and replicating commands by computing
retention from the first addressable segment rather than from the first
segment in the manifest.