datalake/verifier: add ability to handle transactions with gaps
Currently verifier expects to be able to consume/query until hwm. This
does not hold true for transactions due to the offset gaps created by
control batches (commit/batches) and aborted data batches.

A simple example is when the topic data ends with an aborted transaction
resulting in a range of offsets being non consumable.

This commit changes the invariants that mark completion of translation
to the following

1. All offsets until hwm should be consumed
2. All consumed messages must be translated
3. Max consumption offset should be same max translated offset

(1) is where most of the change is as we cannot rely on message offsets
to determine if the consumption is complete. Eg: the topic ends with an
aborted transaction (which includes 1 or more aborted batches and a
control batch that is not propagated to the app layer)