memory_groups: Introduce reserve for partitions
Static partition memory use is currently unaccounted for in memory
groups.

This patch introduces a reserve to account for partition memory usage.

We don't use a share like some of the other groups but rather an upfront
percentage based. This is because we don't want the partitions share to
shrink once we add more groups.