ct: Add circuit_breaker utility
The circuit breaker is a mechanism which can throttle the workload if
cirtain amount of errors were observed. The circuit_breaker (CB) is
created in 'open' state. In this state the requests are supposed to be
handled normally. It has two parameters, error_capacity and cooldown
period. The error_capacity defines how many errors the CB has to observe
to transition into the closed state. In the closed state the requests
are dropped. After cooldown interval the CB transitions from 'closed'
into 'half_open' state. In this state the requests are throttled but not
completely. Some requests are handled normally and some are discarded.
After the next cooldown period the CB transitions back to 'open' state
if there were no errors while requests have been throttled. If any error
occured during 'half_open' stage the CB transitions back to 'closed'
state for the next 'cooldown' interval.

If the errors are registered infrequently (less than one error per
cooldown interval) the CB will not transition into the closed state. So
basically every registered error has a TTL which is equal to cooldown
interval. Without this the CB will be sensitive to sporadic low
frequency noise which we always have in production (cloud "weather").
Eventually, even if errors are very rare the CB will get closed. To
avoid this we have this TTL mechanism that discards error if they're not
frequent enough.

The CB doesn't throttle requests on its own. It provides two methods:
- 'register_error' registers information about the error (timestamp,
  count) and handles state transitions.
- 'state' returns current state of the CB but also handles state
  transitions.
The user of the CB is supposed to record errors and use 'status' method
to decide if the requests should be throttled or blocked. It's up to the
caller to decide how much throttling should be applied. The amount of
throttling should be enough to prevent overload.

One potential usage example is cache service. If the cache is blocked
because of the trim we will get timeout errors from it. The connections
may pipe up if the throttling is in progress. The CB could be used to
block new connections from being created and throttle the requests afrer
the cooldown. This will slow down connection pipeup process.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>