iobuf: update the allocator schedule
iobuf uses a growing schedule of allocations (growth factor 1.5) for
each fragment. This schedule was not aware of the seastar allocation
sizes, so it would request, for example 66K allocation, but seastar
internally rounds that up to 128K, so for that allocation we waste
~50% of the memory which is invisible to iobuf.

In this change we update the schedule to be seastar allocator aware,
i.e., using the same ~1.5x growth factor, but rounding up to the next
seastar allocator boundary. At 16K or below, these boundaries are
log-linear: every 2^n size, plus 3 evenly spread sizes in between
each power of 2. Above 16K, sizes are 2^n.

This change slightly reduces the total number of steps until we get
to the max size of 128K, as from 32K to 128K we use doubling steps
instead of 1.5. For large iobufs this results in 1 fewer total
which is why some tests which tested the exact number of fragments
needed to be decreased by 1.

Fixes CORE-8478.