raft/conensus: change the replicate method to accept vector
Previously `conesnsus::replicate` accepted `model::record_batch_reader`
as an argument. This is a bit of an overuse of reader as every time
batches are replicated the code simply wraps a vector of batches with
in memory record batch reader. This commit simplifies the interface and
changes the interface to accept `chunked_vector<model::record_batch>`

Signed-off-by: Michał Maślanka <michal@redpanda.com>