test: redirect boost test output under bazel
The bazel test runner sets an XML_OUTPUT_FILE environment variable so
that the test writes the results in a local that bazel can subsequently
consume. If the test does not write anything to that location bazel
instead builds a "default" result file which treats the entire test run
as a single test case, and uses the stdout of the test binary as the
output of the test case. In addition to not exposing the test cases
contained in the test correctly, this process is extremely slow when
there is a lot of test output (as many redpanda fixture tests have)
since the output is injected via shell variable interpolation which
apparently doesn't scale up to variables several GB in size.

This change introduces a boost global test configuration object which
looks for the XML_OUTPUT_FILE variable and if set enables JUNIT output
redirected to the specified path.

Fixes CORE-8013.