`storage`: use `model::next_offset()` in `segment::force_roll()`
If the dirty offset is default initialized to `int64_t::min()`
and not updated before a segment is force rolled, an assert will fail
in `disk_log_impl::new_segment()` for the offset being < 0.

Use `model::next_offset()` instead of simply adding `1` to the dirty offset
to avoid this case.