controller: expose recovery state from controller_api
Exposed the recovery state from controller_api, the recovery state is a
metric providing insight into the process of follower recovery. The
state returned from recovery state is based on the raft leader follower
state. This makes the state most up to date and accurate. This will
allow us to present more up to date information to the users. Currently
we are using health reports which is delayed and may be inaccurate.

Signed-off-by: Michał Maślanka <michal@redpanda.com>