`storage`: fix `key_offset_map::initialize()` logic
This would previously overshoot the `size_bytes` provided to it
by filling with `elements_per_fragment()` at least once.

In the lower limit, when `required_entries` is less than `elements_per_fragment()`,
we should be taking the minimum of the two values and pushing back that
number of objects to the `entries` container.