cloud_io: Treat 'unknown' backend as a valid case
The switch stmnt in the 'remote' class triggers 'unreachable' error in
case if the backend is unknonw. This is not valid because the backend
type is detected from the configuration and we shouldn't crash because
configuration is unexpected. Also, it's safe to assume that many cloud
storage providers which implement S3 compatible REST API will be
detected as 'unknown' and should be treated as S3 compactible.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>