datalake: opt out of stop_translator when stopping
Previously stop_translator() seemed to allow erased translators to go
un-stopped if the datalake_manager's gate is closed. This commit adds an
early return to avoid this.

It's not clear that this is actually problematic in practice because the
notification deregistration and gate close should happen atomically in
stop(), so it seems like notifications _should_ be guaranteed to run
without thinking about the gate. But it's an easy enough check to add
in.

(cherry picked from commit 44354fb2dd9e61c323b520393f4d6a2ecf22f230)