storage: Fallback mechanism for index anomalies
The compressed index requires offsets and file offsets to be monotonic.
It also requires the file positions to be separated by 32KiB or more. In
case if some index on disk doesn't fit this requirement the assertion
could be triggered. To avoid this this commit adds a fallback mechanism.
When the data is read from the iobuf we first check if the range is
monotonic. If this is not the case we're using the original index
structure. If everything is OK we're using the compressed index.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>