fmt: remove formatter workaround for old fmt
We had a workaround for fmt versions less than 9.0.1 to avoid sstring
being formatted using operator<< and to use the formatter<>
specialization instead, but this is not needed after version 9.0.1, and
it so was protected by #if FMT_VERSION < 090100. We can just remove
this case entirely on dev as we have upgraded format to 9.0.1 on both
the bazel and cmake sides.

We leave in place only a check that the FMT version is high enough.