cloud_storage: Introduce non-fatal metadata anomalies
The delta-offset anomalies are not fixable becuase the ntp_archiver is
getting delta values from the storage/offset-translator. If the value
which it's getting differs this mean that subsequent retries will yield
the same delta value. If we will not accept the upload the TS will get
stuck and data will not be evicted from the local disk until we will
lose availability.

The alternative is to allow metadata with offset-delta anomalies and
trigger an error message to notify the user.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>