`datalake`: add `translation/utils.cc`
And most importantly, add the function `get_translated_log_offset()`.
This function will be used to compute the appropriate highest translated
log offset for a given translated kafka offset while taking into account
translator batches.

This is an important function for removing offset translation from within
the `translation_stm`, and to be less pessimistic about the
`max_collectible_offset` returned by the `stm` in the future.