bazel/compilation_database_generator: remap compiler paths
For toolchains llvm, the compiler is a shell script that just points to
the actual directory containing the LLVM toolchain. clangd uses the
directory of the first arg (assumed to be the real compiler, not a shell
wrapper) for search paths of stdlib headers. If we remap that then
clangd can find libc++ headers and we can then resolve the headers for
our Bazel provided compiler 🎉