rptest/s: always cleanup azurite containers to avoid slowdown
In https://github.com/redpanda-data/redpanda/pull/24550 a test was
enabled to also run with Azure/Azurite cloud storage.

Azurite is a docker container shared across all tests of a ducktape run.
It is a single threaded nodejs service. It seems that its performance
degrades when buckets aren't removed and some tests become flaky as a
result.

We have somewhat reasonable time budgets for test operations but Azurite
becomes unreasonable slow.

To overcome this we always cleanup Azurite buckets at end of tests. I
did run release tests with and without cleanup and didn't observe
significant change in the runtime of the test suite.

Minio (s3 api equivalent) doesn't seem to have this problem.

The check now will also be skipped for real Azure cloud too. Deleting a
bucket does not require emptying it first so it is a cheap operation.