r/utils: fixed setting log start delta in snapshot
When topic partition is recovered from cloud data Raft protocol is
seeded with special snapshot created during recovery. The snapshot may
then be used to seed out of date followers. The snapshot content must
be valid in terms of offsets and offset translator deltas.

Previously the delta set in the snapshot created during recovery was set
to 0. When this snapshot is delivered to the follower it will reset the
follower offset translation state accordingly. This may lead to
inconsistent deltas across replicas.

Signed-off-by: Michał Maślanka <michal@redpanda.com>