`storage`: fix race between `new_segment()` and `close()`
`disk_log_impl::close()` and `disk_log_impl::remove()` both set
`_closed` to `true`. There is an assert in `new_segment()` which checks
that the log is not closed under `_segments_rolling_lock`. However,
neither `close()` or `remove()` previously respected this lock, which
could lead to a race condition and a triggered assert,
`"cannot add log segment to closed log"`.

Await the `_segments_rolling_lock` in `close()` and `remove()`, and
additional indicate it as `broken()` for future waiters before setting
`_closed = true` to avoid this race.