ducktape: Respect rpk timeout in rpk
Pass the timeout value for rpk that is passed to subprocess also to rpk
itself via `request_timeout_overhead`.

Otherwise the specified timeout will not really be respected if the
default timeout in rpk for a certain command is shorter.

Because `request_timeout_overhead` is on top of other possible timeouts
the total timeout might be larger than the subprocess one but that
should be fine.

(cherry picked from commit d1accb1b12334549dfb6d53b4f6116246b7f4ea3)