ct: Refactor pipeline interface
Introduce 'base_pipeline' that acts as a container which can be used by
both read and write pipelines.

The 'read_pipeline' is derived from 'base_pipeline'. The pipeline
components that use read_pipeline are no longer have to have a reference
to pipeline itself. They can interact with the pipeline through the
'stage' component that automates event subscription.

The 'read_pipeline' users now don't have to access the pipeline and can
just use the corresponding pipeline stage. The creation of the object
hierarchy is also becomes simpler because of this.

The 'write_pipeline' will be modified in the next commit.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>