datalake: lift table_id generation from table_creator
Motivation for this is DLQ project in which we need to manage multiple
tables per topic. To make this possible, we change the public interface
of the table_creator and the RPC interface to coordinator to use
table_identifier as the name for the table rather than topic. Topic name
is kept in the interface because it has some other uses than just table
name generation.

Note that this is a breaking change in the RPC protocol so creating
tables/committing translated files will stall until the cluster is fully
upgraded. This should be acceptable for pre-GA.

Rejected alternatives:
- Custom RPC operation for DLQ table: in the future we will likely have
  to manage multiple tables for breaking schema changes. The approach
  with lifting table_id should make that possible too.
- Bool `is_dlq_table` flag: rejected for the same reason as above.