cluster: nested partition probe infrastructure
There are components which need to report metrics but their lifetime
might be shorter than what we want for metrics. We solve this problem
with partition_probe_part. With it we can handle probes to the
partition_probe and let it manage them and their lifetimes. Users can
get a reference to the probe and update it as needed.

Lazy initialized too.