bazel: extract //src/v/cluster:partition_probe
Unfortunately we still depend on cluster... With the infrastructure
introduced in previous commit we should be able to easily decouple
partition_probe from cluster btw.

Scrubber and schema registry could have their nested_partition_probe.
Out of the box I believe there is a potential problem with duplicated
group names so the API might need a bit of massaging.