sr: Perform raw schema lookup first
Certain SR clients perform the following pattern to lookup schema IDs:

* POST /subject/sub/versions?normalize=true -d {schema}
* GET /subject/sub/versions/latest -> this returns the schema and latest
  version but NOT IDs
* POST /subject/sub -d {schema-from-get} -> this returns ID

In our SR, a call to POST /subject/sub without normalize=true will
result in this schema being sanitized which may change the content of
the schema.  The change introduced in this PR is to perform a direct
lookup of the raw provided schema before we either normalize or sanitize
it.

Signed-off-by: Michael Boquard <michael@redpanda.com>