ct: Implement dl_stm snapshots
Implement local snapshots and Raft snapshots.
Local snapshots are simple. The snapshot is basically a serialized
dl_stm_state instances.

The Raft snapshots are a bit more complicated. Because the Raft snapshot
is taken at offset we can't serialize dl_stm_state. The state might
contain data added by commands whith offsets larger than the one used by
Raft. To fix this the copy of thte state is created. In this copy all
versions that exceed the requested Raft offset are removed and the
offsets are adjusted accordingly. The assumption is that all commands
which are not included into the snapshot will be replayed by the stm
manager when the snapshot will be installed.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>