`config`: move `validate_config` to `config::configuration`
Moves the existing multi-property validation portion of the function
from `admin/server.cc` into `config::configuration` as a static member
function. The rest of the existing function which validate schema registry
and admin endpoint users is left in `admin/server.cc`.

This encapsulates the logic better, and also allows it to be re-used in
other parts of the code which need to perform validation on cluster configs.

Also, change in/out `std::map` parameters to a return struct, containing
a `set` of properties that need to be unset in case of a bad configuration,
as well as the `map` of properties and their corresponding errors.