`storage`: rewrite compressed batches with `compression.type` during compaction
The broker/topic properties `log_compaction_type`/`compression.type` are not
currently used for broker side compression in `redpanda`. The only compression
`redpanda` supports is producer side compression, and we make no attempts to
respect the broker/topic properties.

This commit is half the battle. Now, when re-writing raft data batches in segments
during compaction, we will respect `compression.type` and actively re-compress
segments according to this property. If the `compression.type=producer`, we
respect the original compression type in the record batch's header, and
re-compress using that. If the `compression.type=none`, we will not re-compress
the batch, even if it was compressed on the producer side originally. If
`compression.type` is any other specific compression algorithm, the batch is
re-compressed using the provided algorithm.

The other half of the battle is broker compression on the produce path- this is
left as future work.