base: avoid atomics inside the vassert macro
Produce microbenchmarks have gotten slower since the latest change to
vassert, which we suspect is caused by the introduction of atomic
operations on the (unlikely) branch taken in the case of a vassert
crash.

This moves the implementation closer to the original implementation by
using a non-atomic function pointer for the callback. This is safe to do
because we now register the callback before any threads (reactor or
non-reactor) are created, which ensures that the callback is visible on
all threads. This avoids the data race that would otherwise exist if the
callback was registered while other threads are already running.