`rptest`: increase `initial.broker.registration.timeout.ms` from default
This test fails often due to a broker failing to register in time:

```
[2025-03-12 22:50:14,012] ERROR [BrokerLifecycleManager id=1] Shutting down because we were unable to register with the controller quorum. (kafka.server.BrokerLifecycleManager)
```

It may be the case that the broker just needs more time than the default
value of 60 seconds to be successful in joining the controller quorum.

Bump the timeout by setting it as a `server_prop_override`.

Also add a timeout parameter to `KafkaServiceAdapter.start()` to match.